requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
email-validator==1.0.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-Select2==0.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
Pillow==7.1.1
psycopg2==2.8.5
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-slugify==4.0.0
six==1.14.0
SQLAlchemy==1.3.16
text-unidecode==1.3
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
